+++ vpin-studio-server/src/main/java/de/mephisto/vpin/server/games/GameService.java
@@ -275,7 +275,6 @@ public class GameService implements InitializingBean {
         }
       }
 
-
       if (descriptor.isDeleteAltColor()) {
         if (game.getAltColorFolder() != null && !FileUtils.deleteFolder(game.getAltColorFolder())) {
           success = false;