+++ RELEASE_NOTES.md
@@ -10,7 +10,10 @@
   <img src="https://raw.githubusercontent.com/syd711/vpin-studio/main/documentation/preferences/volume-control.png" width="400" />
 
 - **Table Validators**: Added validator for outdated recordings. //TODO check date
-- **Table Management / Backglass Section**: Added "Delete" button for backglasses.
+- **Table Management / Backglasses**: Added "Delete" button.
+- **Table Management / ALT Sound**: Added "Delete" button.
+- **Table Management / ALT Color**: Added "Delete" button.
+- **Table Management / PUP Packs**: Added "Delete" button.
 
 ### Bugfixes